refactor: move fetch_table/s to table_routes.rs
